create selectedEntityView
